moved db_dummy.py and dummy.json into population/
